Set break (#3)
* scannerc: Added support for CRLF edge case

There is an edge case where the scanner peeks ahead and finds an LF byte,
and thinking it's in a valid line break position it appends a \n to the
text. However, this ignores the possibility that this is actually the LF
of a complete CRLF.

* scanner: remove leftover println

* Add support for setting line break style

The ability to set line break style for the emitter under the hood
exists already in `apic.go`. This PR adds an API to set the line break
style on the Encoder.

* Add godoc for SetLineBreakStyle

* adjust test to remove magic number