encode: add SetExplicitDocumentStart option
Adds an option to the emitter to always have an explicit document start.
This option can be used as an override to the default behaviour, which
checks that it's not the first document start or that it's canonical
style before setting something as explicit. This allows for the opening
document start to always be written (even when it wasn't originally
specified).